Fix strncat warning.
The warning is that the code is copying a four character string
PLUS a NUL character into a buffer of size four.  This isn't possible
without overflowing the buffer.  Since the intent seemes to be to
copy the string "NULL" and null terminate it, switch to a simple call
to strcpy.